requirements.txt
mysql-connector-python~=8.0.21
colorama~=0.4.3
setuptools~=50.3.0